New cygwin setup: download cygwin cache and offline install later
The previous setup was ideal in two regards.

1. It doesnt bother the user with Cygwin installation window.
2. Reproducible: every esy-bash user at a certain version would always
get the same version of cygwin packages. This was a huge win in terms
of overall reproducibility of esy sandboxes.

However, recently, it has become difficult to push to NPM due to size
of the tarball - NPM registry stops accepting them. Publishing them to
Github's registry wasn't viable because at them time of this writing,
they need Github Personal Access Tokens just to install it (even if
the package is public)

To work around this, the we compromise with #1 but preserve #2.

We download all the cygwin packages on the CI, store them at
.cygwin/var/cache/setup, and later run the Cygwin installer asking it
to work in offline mode, ie, from local cache at /var/cache/setup.

This way,
1. postinstall doesn't hit network
2. users get the same version of cygwin packages.

Unfortunately, with Cygwin Window will show up. Perhaps we could work
around this later.